+++ shared/src/commonMain/kotlin/de/bissell/starcruiser/Commands.kt
@@ -28,7 +28,7 @@ sealed class Command {
     data class CommandChangeThrottle(val value: Int) : Command()
 
     @Serializable
-    data class CommandChangeJumpDistance(val value: Int) : Command()
+    data class CommandChangeJumpDistance(val value: Double) : Command()
 
     @Serializable
     data class CommandChangeRudder(val value: Int) : Command()
